Move hasher to common
