+++ gateways/utils.go
@@ -24,13 +24,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
-// Hasher hashes a string
-func Hasher(value string) string {
-	h := fnv.New32a()
-	_, _ = h.Write([]byte(value))
-	return fmt.Sprintf("%v", h.Sum32())
-}
-
 // SetValidateReason set the result of event source validation
 func SetValidEventSource(v *ValidEventSource, reason string, valid bool) {
 	v.Reason = reason
